Reorganized code for clarity and cleaned up the imports and exports
